Added new file to start bots up
Created robotStarter.py, a file which will deal with loading and unloading encodings into the bots. It will read the encoding and generation number, save those, run the encodings and get the evaluations which it will then save to a new file for the next generation of bots. Also added more methods for the encoding class